Add area chart. Get rid of removeAll from chart class. Updated colors. Better but not perfect. See description.
Added @seriesCanvas() which contains g node with all of the elements
for this particular series. Without it each renderer could mess with
all the stuff in the graph. This was leading to a small chaos and
adding many lines to a single chart was too painful. You would need to
keep an id in each of the paths or circles etc.
One last thing is that I needed to add remove() to the
draggable_line_renderer. Hope to fix it later.