References #31 Don't check y0 about being a number when filtering the data or rendering columns
If we have two series and one of them has undefined vaule for, say, "Y"
then the Y0 will be undefined even if the currently rendered series
values are correct. This is because Y0 is a sum of all Y's at the given
X position. It will fail because you cannot add undefined to a number
and therefore, Y0 will be undefined.